Tried merging all the files
