+++ subete/__init__.py
@@ -1,6 +1,3 @@
-from .project import *
-from .sample_program import *
-from .language_collection import *
 from .repo import *
